Refactored authentication
takes place solely in api/index.js then adds a user property to the request if a valid token is passed in
