+++ api/index.js
@@ -1,8 +1,13 @@
 const express = require('express');
 const cors = require('cors')
 const router = express.Router();
+const jwt = require('jsonwebtoken')
+const { JWT_SECRET } = process.env;
 router.use(cors())
 
+const {
+    getUserById
+} = require ('../db')
 
 // GET /api/health
 router.get('/health', async (req, res, next) => {
@@ -12,6 +17,33 @@ router.get('/health', async (req, res, next) => {
     })
     next()
 });
+
+router.use(async (req, res, next) => {
+    const prefix = 'Bearer ';
+    const auth = req.header('Authorization');
+  
+    if (!auth) { 
+      next();
+    } else if (auth.startsWith(prefix)) {
+      const token = auth.slice(prefix.length);
+  
+      try {
+        const { id } = jwt.verify(token, JWT_SECRET);
+  
+        if (id) {
+          req.user = await getUserById(id);
+          next();
+        }
+      } catch ({ name, message }) {
+        next({ name, message });
+      }
+    } else {
+      next({
+        name: 'AuthorizationHeaderError',
+        message: `Authorization token must start with ${ prefix }`
+      });
+    }
+  });
 
 // ROUTER: /api/users
 const usersRouter = require('./users');
